Global Health Expenditure Analysis . Read me
Project Overview:
In this project, I have conducted a comprehensive analysis of global health expenditure, with a specific focus on the United States of America. The primary objective was to gain valuable insights into healthcare spending trends and allocations across countries and regions. The key finding of this analysis is that the USA allocates approximately 45% of its Gross Domestic Product (GDP) to health expenditure, highlighting the significant financial commitment to healthcare in the country.